+++ src/cli.py
@@ -6,6 +6,7 @@ from handlers import (
     check_for_updates_handler,
     fetch_handler,
     list_settings_handler,
+    open_url_handler,
     search_handler,
     totp_handler,
 )
@@ -17,12 +18,13 @@ class CLIActions:
     SETTINGS_LIST = "settings_list"
     TOTP = "totp"
     CHECK_FOR_UPDATES = "check_for_updates"
+    OPEN_URL = "open_url"
 
     @classmethod
     def choices(cls) -> t.List[str]:
         """Returns an action list."""
 
-        return [cls.SEARCH, cls.FETCH, cls.SETTINGS_LIST, cls.TOTP, cls.CHECK_FOR_UPDATES]
+        return [cls.SEARCH, cls.FETCH, cls.SETTINGS_LIST, cls.TOTP, cls.CHECK_FOR_UPDATES, cls.OPEN_URL]
 
 
 def parse_args() -> argparse.Namespace:
@@ -46,6 +48,10 @@ def parse_args() -> argparse.Namespace:
 
     check_for_updates_parser = subparsers.add_parser(CLIActions.CHECK_FOR_UPDATES)
     check_for_updates_parser.set_defaults(handler=check_for_updates_handler)
+
+    open_url_parser = subparsers.add_parser(CLIActions.OPEN_URL)
+    open_url_parser.set_defaults(handler=open_url_handler)
+    open_url_parser.add_argument("url")
 
     return parser.parse_args()