+++ tests/src/cli/test_functions.py
@@ -31,5 +31,6 @@ class TestParseArgs:
         add_parser_mock.assert_any_call(CLIActions.FETCH)
         add_parser_mock.assert_any_call(CLIActions.TOTP)
         add_parser_mock.assert_any_call(CLIActions.CHECK_FOR_UPDATES)
+        add_parser_mock.assert_any_call(CLIActions.OPEN_URL)
 
-        assert add_parser_mock.call_count == 5
+        assert add_parser_mock.call_count == 6