+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clipboard timeout.
 - Show note details in full text with `Opt ⌥ + Return ↵`.
 - Request to check for workflow updates.
+- Open url in default browser with `Opt ⌥ + Return ↵`
 
 ## [2.1.0] - 2022-04-10